making slot properties optional and checking for default values
