+++ src/types/ability.ts
@@ -1,5 +1,6 @@
-import { MessageData, ValidateResult, WolfState } from './state'
-import { SetSlotDataFunctions, GetSlotDataFunctions, GetStateFunctions, SlotConfirmationFunctions } from './function'
+import { WolfState } from './state'
+import { GetStateFunctions } from './function'
+import { Slot } from './slot';
 
 /**
  * Defines conversation abilities, used to control overall flow
@@ -21,34 +22,4 @@ export interface Ability<T> {
 export interface NextAbilityResult {
   abilityName: string,
   message: string | null
-}
-
-/**
- * Wolf primitive representing data points that should be collected.
- */
-export interface Slot<T> {
-  name: string,
-  defaultIsEnabled?: boolean,
-  order?: number,
-  query: (convoState: T, getSlotDataFunctions: GetSlotDataFunctions) => string,
-  validate: (submittedValue: any, messageData: MessageData) => ValidateResult,
-  retry: (convoState: T, submittedData: any, turnCount: number) => string,
-  onFill: (
-    submittedValue: any,
-    convoState: T,
-    setOtherSlotFunctions: SetSlotDataFunctions,
-    confirmationFunctions: SlotConfirmationFunctions
-  ) => string | void
-}
-
-export interface ShouldRunCompleteResult {
-  shouldComplete: boolean,
-  reason?: string,
-  nextAbility?: string
-}
-
-export interface IncomingSlotData {
-  slotName: string,
-  abilityName: string,
-  value: string
 }
