+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './function'
 export * from './state'
 export * from './generic'
 export * from './storage'
+export * from './slot'
